+++ src/States/RaceModes/RaceStatesBase.cpp
@@ -73,7 +73,7 @@ void RaceStatesBase::InitMap()
 void RaceStatesBase::InitPlayerLoc()
 {
     std::string name = m_data->user.getMapGame().substr(0,m_data->user.getMapGame().find('.'));
-    auto locations = readFromFile<int>(name+"_start_position.txt",3,2);
+    auto locations = readFromFile<float>(name+"_start_position.txt",3,2);
     m_player.setFinishLine(locations[2][0]);
 		m_player.setLocation(sf::Vector2f(locations[0][0],locations[0][1]));
